refactor: extract Ctx into separate module
Move the Ctx struct and its impl into checker/ctx.rs for better organization.
Re-export Ctx from the checker module to maintain the public API.
